refactor(cqm): Merge DataElement into QDMBaseType
Remove unnecessary inheritance layer by merging DataElement's
functionality directly into QDMBaseType. This fixes the _type
property type conflict and simplifies the class hierarchy.

- Move array constructor, addCode(), dataElementCodes to QDMBaseType
- Have QDMBaseType implement JsonSerializable directly
- Delete DataElement.php
- Update Generator to skip QDMBaseType (now manually maintained)
