+++ src/Cqm/Qdm/QDMBaseType.php
@@ -2,18 +2,28 @@
 
 namespace OpenEMR\Cqm\Qdm;
 
+use Exception;
+use JsonSerializable;
+use OpenEMR\Cqm\Qdm\BaseTypes\Code;
+
 /**
  * OpenEMR\Cqm\Qdm\QDMBaseType
  *
- * This is a class generated with Laminas\Code\Generator.
+ * Base class for generated QDM types. Provides array-based construction
+ * for hydrating from external data.
  *
  * @QDM Version 5.6
  * @author Ken Chapple <ken@mi-squared.com>
  * @license https://github.com/openemr/openemr/blob/master/LICENSE GNU General
  * Public License 3
  */
-class QDMBaseType extends \OpenEMR\Cqm\Qdm\BaseTypes\DataElement
+class QDMBaseType implements JsonSerializable
 {
+    public mixed $_type = 'QDM::QDMBaseType';
+    public ?string $bundleId = null;
+    /** @var list<Code> */
+    public array $dataElementCodes = [];
+
     /**
      * @property string|null For backwards compatibility
      */
@@ -39,5 +49,27 @@ class QDMBaseType extends \OpenEMR\Cqm\Qdm\BaseTypes\DataElement
      */
     public $qdmVersion = '5.6';
 
-    public $_type = 'QDM::QDMBaseType';
+    /**
+     * @param array<string, mixed> $properties
+     */
+    public function __construct(array $properties = [])
+    {
+        foreach ($properties as $property => $value) {
+            if (property_exists($this, $property)) {
+                $this->{$property} = $value;
+            } else {
+                throw new Exception("Property {$property} does not exist on " . static::class);
+            }
+        }
+    }
+
+    public function addCode(Code $code): void
+    {
+        $this->dataElementCodes[] = $code;
+    }
+
+    public function jsonSerialize(): mixed
+    {
+        return get_object_vars($this);
+    }
 }
